Refactor Feed component in HomePage.tsx into its own Feed component
